+++ README.md
@@ -1,2 +1,3 @@
 To install, do the pip thing with requirements.txt then python the
 server. Have fun.
+Also requirements.txt is wayyyyy out of date
