--- colab/README.md
@@ -1,3 +1,0 @@
-Here's how to test the server:
-curl -H 'Content-Type: application/json' -X POST --data @example.json https://c1102ab6.ngrok.io/webhook
-but replace ngrok with wherever the server really is.
