--- colab/curl_webhook.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-curl --insecure -X POST -H "Content-Type: application/json" --data @example.json https://35.236.232.76:5000/webhook
-
