+++ test_webhook.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+curl --insecure -X POST -H "Content-Type: application/json" --data @example_webhook_call.json https://sleepstorymachine.xyz:5000/webhook
+
